--- src/STW.ProcessingApi.Function/Models/Commodity/CommodityAttribute.cs
@@ -1,8 +1,0 @@
-namespace STW.ProcessingApi.Function.Models.Commodity;
-
-public class CommodityAttribute
-{
-    public string CommodityCode { get; set; }
-
-    public string Propagation { get; set; }
-}
